separate interfaces for all objects in the app
