+++ src/components/ui/sheet.tsx
@@ -126,6 +126,6 @@ SheetDescription.displayName = SheetPrimitive.Description.displayName
 
 export {
   Sheet, SheetClose,
-  SheetContent, SheetDescription, SheetFooter, SheetHeader, SheetOverlay, SheetPortal, SheetTitle, SheetTrigger
+  SheetContent, SheetDescription, SheetFooter, SheetHeader, SheetOverlay, SheetPortal, SheetTitle, SheetTrigger, sheetVariants
 }
