+++ src/interfaces/badgeProp.tsx
@@ -1,0 +1,5 @@
+import {type VariantProps } from "class-variance-authority"
+import { badgeVariants } from "@/components/ui/badge"
+export interface BadgeProps
+  extends React.HTMLAttributes<HTMLDivElement>,
+    VariantProps<typeof badgeVariants> {}
