+++ src/interfaces/buttonProp.tsx
@@ -1,0 +1,15 @@
+import { type VariantProps } from "class-variance-authority"
+import { buttonVariants } from "@/components/ui/button"
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean
+}
+
+export interface AnimatedButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  variant?: 'primary' | 'secondary' | 'outline';
+  size?: 'default' | 'sm' | 'lg' | 'icon';
+  children: React.ReactNode;
+  className?: string;
+}
