+++ src/interfaces/chartDataPoint.tsx
@@ -1,0 +1,8 @@
+export interface ChartDataPoint {
+    date: number;
+    formattedDate: string;
+    weight: number;
+    oneRepMax?: number;
+    maxReps?: number;
+    maxVolume?: number;
+  }
