+++ src/interfaces/commandDial.tsx
@@ -1,0 +1,2 @@
+import { type DialogProps } from "@radix-ui/react-dialog"
+export interface CommandDialogProps extends DialogProps {}
