+++ src/interfaces/locationState.tsx
@@ -1,0 +1,3 @@
+export interface LocationState {
+    from?: string;
+  }
