+++ src/interfaces/textArea.tsx
@@ -1,0 +1,2 @@
+export interface TextareaProps
+  extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {}
